Add `Study`, `Model`, `Sequence`, `Group` classes
`Study` is a container class that coordinates the parsing of wresl
files, while tracking the configuration settings that will be need to
compile those files. The first implementation does not compile the
wresl files, but simply collects the ParseTree objects using the
`ExpandIncludesVisitor` object to do that.

`Model`, `Sequence`, and `Group` are all stubbed classes, that will be
implemented to support the compilation process of the `ParseTrees`